src/cmd update
